backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.10.6
pydantic-settings==2.8.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
openai==0.28.0
boto3==1.28.38
alembic==1.12.0
pytest==7.4.2
httpx==0.24.1
pymongo==4.5.0
python-dotenv==1.0.0
pydub==0.25.1
openai-whisper==20240930
email-validator==2.2.0
PyPDF2==3.0.1
